binutils: fix .init/.fini literals moving in xtensa gas
Despite the documentation and the comment in xtensa_move_literals, in
the presence of --text-section-literals and --auto-litpools literals are
moved from the separate literal sections into .init and .fini, because
the check in the xtensa_move_literals is incorrect.

This fixes build errors seen with projects that have .init/.fini and use
text-section-literals.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>