Remove -v from strip_args
FreeBSD base strip (which is not Binutils strip but elfutils
strip) does not not handle -v.

Rather than creating a separate branch for CT_HOST *freebsd* or
merge it into *darwin*, just remove -v altogether. The output along
the lines of

  copy from `foo' [elf64-x86-64] to `stx15vcK' [elf64-x86-64]

does not bear any real information (it's too much for a progress
bar, and too little for anything else).